fagicrm/requirements.txt
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.1
django-extensions==3.2.3
python-decouple==3.8
whitenoise==6.6.0
gunicorn==21.2.0
mysqlclient==2.2.0